blob_function_gcp/requirements.txt
functions-framework==3.8.1
google-cloud-storage==2.16.0
